Add printf implementation
I will get back on to this, there is a part of it where I can't fully
understand. Where if the *fmt is x, it uses the most significant
4bits to the least significant. 1byte = 8bits, nibble = 4bits, or half a
byte.

According to my research, humans tend to read left to right, while
computers tend to read it as right to left. This is because of
endianess.